requirements.txt
django==5.2.8
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.29.0
drf-spectacular-sidecar==2025.10.1